The block-inverse is working now.
In the next step the more refined version considering the off-diagonal
blocks will be added. A clear improvement of convergence-speed compared
to GMRES with identity-preconditioning is already visible (convergence
in finite time i.e. 650 steps for 40k dofs. This will likely be improved
upon with the afore mentioned additions.)

The program can currently be compiled and executed.